refactor: move TestRequestButton to a new feature folder
