+++ packages/api-reference/src/components/Content/Webhooks/index.ts
@@ -1,1 +1,1 @@
-export { default as Webhooks } from './Webhooks.vue'
+export { default as Webhooks } from '../../../features/Operation/components/Webhooks.vue'
