+++ packages/api-reference/src/embeds/ExampleRequest/ExampleRequest.vue
@@ -2,7 +2,7 @@
 import { useServerStore } from '#legacy'
 import { ExampleRequest } from '@/features/ExampleRequest'
 import OperationPath from '@/features/Operation/OperationPath.vue'
-import TestRequestButton from '@/features/Operation/TestRequestButton.vue'
+import { TestRequestButton } from '@/features/TestRequestButton'
 import type { Spec, TransformedOperation } from '@scalar/types/legacy'
 import { toRef } from 'vue'
