+++ packages/api-reference/src/features/Operation/index.ts
@@ -1,1 +1,2 @@
 export { default as Operation } from './components/Operation.vue'
+export { default as Webhooks } from './components/Webhooks.vue'
