+++ packages/api-reference/src/features/Operation/layouts/ModernLayout.vue
@@ -10,13 +10,13 @@ import {
 } from '@/components/Section'
 import { ExampleRequest } from '@/features/ExampleRequest'
 import { ExampleResponses } from '@/features/ExampleResponses'
+import { TestRequestButton } from '@/features/TestRequestButton'
 import type { TransformedOperation } from '@scalar/types/legacy'
 
 import OperationDescription from '../OperationDescription.vue'
 import OperationParameters from '../OperationParameters.vue'
 import OperationPath from '../OperationPath.vue'
 import OperationResponses from '../OperationResponses.vue'
-import TestRequestButton from '../TestRequestButton.vue'
 
 defineProps<{
   id?: string
