+++ packages/api-reference/src/features/TestRequestButton/index.ts
@@ -1,0 +1,1 @@
+export { default as TestRequestButton } from './TestRequestButton.vue'
